fixed points issue
