+++ lib/game.js
@@ -97,7 +97,7 @@ Game.prototype.endGame = function(gameLoop, context, canvas, hearts) {
   if (endingFrames < 25) {
     requestAnimationFrame(gameLoop);
   } else {
-    scoreboard.showHighScoreEntry();
+    scoreboard.showHighScoreEntry(points);
     this.clearCanvas(context, canvas);
     this.showGameOverScreen();
     this.resetGame(hearts);
@@ -122,7 +122,11 @@ Game.prototype.showGameOverScreen = function(){
 
 Game.prototype.sendHighScore = function(){
     scoreboard.sendHighScore(points);
-    points = 0;
+    // points = 0;
+};
+
+Game.prototype.resetPoints = function(){
+  points = 0;
 };
 
 module.exports = Game;
