+++ lib/index.js
@@ -42,6 +42,7 @@ function showCanvas(){
   $("#start-screen").hide();
   $("#canvas-elements").show();
   var startTime = Date.now();
+  game.resetPoints();
   startGame(startTime);
 }
