+++ lib/scoreboard.js
@@ -4,19 +4,20 @@ var firebaseConfig = require('./firebase-config');
 var firebaseApp = firebase.initializeApp(firebaseConfig);
 var fireDb = firebaseApp.database();
 var scoreboardRecords = [];
-const Helpers = require('./helpers');
-var helpers = new Helpers();
+const ScoreboardHelpers = require('./scoreboard-helpers');
+var helpers = new ScoreboardHelpers();
 
 function Scoreboard(){}
 
-Scoreboard.prototype.showHighScoreEntry = function(){
-  $('.scoreboard').hide();
-  $('.high-score-entry').css('display', 'inline-block');
+Scoreboard.prototype.showHighScoreEntry = function(points){
+  if (scoreboardRecords[4].points < points){
+    $('.scoreboard').hide();
+    $('.high-score-entry').css('display', 'inline-block');
+  }
 };
 
 Scoreboard.prototype.sendHighScore = function(points){
     scoreboardRecords = helpers.addToScoreboardRecords(scoreboardRecords, points);
-    // addToScoreboardRecords(points);
     fireDb.ref('highscore/').set({
       highscores: scoreboardRecords
     });
